optimize and hot fix
