+++ src/Components/Workspace/Workspace.jsx
@@ -6,7 +6,6 @@ import SellIcon from "@mui/icons-material/Sell";
 import PlaceIcon from "@mui/icons-material/Place";
 import AccessTimeIcon from "@mui/icons-material/AccessTime";
 import { formatNumber } from "../../Utils/constant.js";
-import { ConstructionOutlined } from "@mui/icons-material";
 
 const Workspace = ({ data }) => {
   const handleClick = () => {
@@ -14,7 +13,7 @@ const Workspace = ({ data }) => {
   };
 
   return (
-    <div className="wrap" onClick={handleClick}>
+    <div className="wrap hover:cursor-pointer" onClick={handleClick}>
       <div className="flex" style={{ gap: "16px" }}>
         <div className="image-container">
           <img src={data.image} className="image" alt="" />
