+++ src/Pages/WorkspaceDetails/WorkspaceDetails.jsx
@@ -2,13 +2,11 @@ import { useState } from "react";
 import { useParams } from "react-router-dom";
 import Navbar from "../../Components/Navbar/Navbar";
 import MenuItem from "@mui/material/MenuItem";
-import Rating from "@mui/material/Rating";
 import FormControl from "@mui/material/FormControl";
 import Select from "@mui/material/Select";
 import CircularProgress from "@mui/material/CircularProgress";
 import Dialog from "@mui/material/Dialog";
 import DialogContent from "@mui/material/DialogContent";
-import Pagination from "@mui/material/Pagination";
 import {
   getWorkspaceDetailsService,
   getWorkspaceReviewsService,
@@ -17,17 +15,13 @@ import { useQuery } from "@tanstack/react-query";
 import { Box } from "@mui/material";
 import iconSet from "../../assets/icon/iconSet";
 import StandardImageList from "../../Components/StandardImageList/StandardImageList";
-import { format, parseISO } from "date-fns";
 
-const convertToProgressBar = (star) => {
-  const progressWidth = Math.round((star / 5) * 320);
-  return progressWidth;
-};
+import ReviewProgressBar from "../../Components/ReviewProgressBar/ReviewProgressBar";
+import Reviews from "../../Components/Reviews/Reviews";
 
 function WorkspaceDetails() {
   const [reviewSortOpt, setReviewSortOpt] = useState(0); // 0: latest, oldest
   const [isPopupPhotoGallery, setIsPopupPhotoGallery] = useState(false);
-  const [currentPage, setCurrentPage] = useState(0);
 
   const workspaceParams = useParams();
 
@@ -59,38 +53,6 @@ function WorkspaceDetails() {
     );
   }
 
-  //
-  const wifi_total_rating = workspaceReviewsData.data.reduce(
-    (acc, item) => acc + Number(item.wifi_rating),
-    0
-  );
-
-  const space_total_rating = workspaceReviewsData.data.reduce(
-    (acc, item) => acc + Number(item.space_rating),
-    0
-  );
-
-  const price_total_rating = workspaceReviewsData.data.reduce(
-    (acc, item) => acc + Number(item.price_rating),
-    0
-  );
-  // workspace details rating average
-  const wifi_average_rating = (
-    wifi_total_rating / workspaceReviewsData.data.length
-  ).toFixed(1);
-
-  const space_average_rating = (
-    space_total_rating / workspaceReviewsData.data.length
-  ).toFixed(1);
-
-  const price_average_rating = (
-    price_total_rating / workspaceReviewsData.data.length
-  ).toFixed(1);
-
-  const spaceReviewClassName = convertToProgressBar(space_average_rating);
-  const wifiReviewClassName = convertToProgressBar(wifi_average_rating);
-  const priceReviewClassName = convertToProgressBar(price_average_rating);
-
   const slicedImage = workspaceDetailsData.data.workspace_images.slice(
     0,
     Math.min(3, workspaceDetailsData.data.workspace_images.length)
@@ -107,29 +69,6 @@ function WorkspaceDetails() {
   const handleChange = (e) => {
     setReviewSortOpt(e.target.value);
   };
-  //handle sort comment
-  if (reviewSortOpt == 0) {
-    workspaceReviewsData.data.sort(
-      (a, b) => new Date(b.created_at) - new Date(a.created_at)
-    );
-  } else {
-    workspaceReviewsData.data.sort(
-      (a, b) => new Date(a.created_at) - new Date(b.created_at)
-    );
-  }
-
-  // Pagination
-  const paginationCount = Math.ceil(workspaceReviewsData.data.length / 10);
-  console.log(paginationCount);
-  const handlePaginationChange = (e, page) => {
-    setCurrentPage(page - 1);
-  };
-
-  const currentReviews = workspaceReviewsData.data.slice(
-    10 * currentPage,
-    Math.min(10 * (currentPage + 1), workspaceReviewsData.data.length)
-  );
-  console.log(currentReviews);
 
   return (
     <div className="text-gray-950">
@@ -169,62 +108,18 @@ function WorkspaceDetails() {
           <div className="w-[599px] h-[362px] p-6 rounded-2xl drop-shadow-sm border-2 border-satoru-blue">
             <h3 className="text-satoru-blue font-bold">Đánh giá</h3>
             <div className="border-[1px] border-gray-700 w-14"></div>
-            <div className="flex justify-center">
-              <div className="flex justify-center items-center w-24 h-24 rounded-lg bg-cyan-500 mr-2">
-                <span className="text-white text-6xl font-semibold">
-                  {workspaceDetailsData.data.average_rating}
-                </span>
+            {workspaceReviewsData.data.length == 0 ? (
+              <div className="w-full h-full flex justify-center items-center">
+                <div className="font-md text-lg w-3/4 h-3/4  rounded flex justify-center items-center  ">
+                  <span>Chưa có đánh giá</span>
+                </div>
               </div>
-              <div className="flex flex-col h-24 justify-evenly">
-                <div className="font-bold text-2xl">Tốt</div>
-                <div className="text-lg">/5 (5 stars)</div>
-              </div>
-            </div>
-            {/* progress bar */}
-            <div className="p-4 flex items-center gap-4">
-              <span className="w-24 font-medium text-base">Không gian</span>
-              <div className="inline-block w-80 h-5 bg-slate-200 rounded-lg">
-                <div
-                  className="  bg-satoru-blue h-5 rounded-lg"
-                  style={{ width: `${spaceReviewClassName}px` }}
-                ></div>
-              </div>
-              <div>
-                <span className="font-medium text-base">
-                  {space_average_rating}/5
-                </span>
-              </div>
-            </div>
-
-            <div className="p-4 flex items-center gap-4">
-              <span className="w-24 font-medium text-base">Wifi</span>
-              <div className="inline-block w-80 h-5 bg-slate-200 rounded-lg">
-                <div
-                  className="  bg-satoru-blue h-5 rounded-lg"
-                  style={{ width: `${wifiReviewClassName}px` }}
-                ></div>
-              </div>
-              <div>
-                <span className="font-medium text-base">
-                  {wifi_average_rating}/5
-                </span>
-              </div>
-            </div>
-
-            <div className="p-4 flex items-center gap-4">
-              <span className="w-24 font-medium text-base">Giá</span>
-              <div className="inline-block w-80 h-5 bg-slate-200 rounded-lg">
-                <div
-                  className="  bg-satoru-blue h-5 rounded-lg"
-                  style={{ width: `${priceReviewClassName}px` }}
-                ></div>
-              </div>
-              <div>
-                <span className="font-medium text-base">
-                  {price_average_rating}/5
-                </span>
-              </div>
-            </div>
+            ) : (
+              <ReviewProgressBar
+                workspaceAvgRating={workspaceDetailsData.data.average_rating}
+                reviews={workspaceReviewsData.data}
+              />
+            )}
           </div>
           {/* Info card */}
           <div className="w-[599px] h-[362px] p-6 rounded-2xl border-2 drop-shadow-sm border-satoru-blue text-lg">
@@ -383,60 +278,28 @@ function WorkspaceDetails() {
           </div>
         </div>
         {/* Review content*/}
-        {/* TODO */}
-        {/* 
-        
-  Nút sắp xếp hoạt động 
-  Sau đó làm pagination sao cho nó có thể chạy được.
-
-        */}
-        {currentReviews.map((review) => (
-          <div className="flex text-left gap-6 my-4">
-            <div>
-              <img
-                src={review.user.avatar_url}
-                alt="avatar"
-                className="w-12 h-12 rounded-full mt-2"
+        {workspaceReviewsData.data.length == 0 ? (
+          <div>Chưa có đánh giá</div>
+        ) : (
+          <Reviews
+            workspaceReviewsData={workspaceReviewsData}
+            reviewSortOpt={reviewSortOpt}
+          />
+        )}
+        <div>
+          <Dialog
+            open={isPopupPhotoGallery}
+            onClose={handleClose}
+            maxWidth={"md"}
+            className="flex justify-center"
+          >
+            <DialogContent>
+              <StandardImageList
+                itemData={workspaceDetailsData.data.workspace_images}
               />
-            </div>
-            <div className="flex flex-col gap-3 w-11/12 bg-slate-100 rounded-lg p-6">
-              <div>
-                <h4 className="font-bold">{review.user.username}</h4>
-                <h5>{format(parseISO(review.created_at), "dd-MM-yyyy")}</h5>
-              </div>
-              <div className="flex items-center gap-4 ">
-                <Rating
-                  name="read-only"
-                  value={review.average_rating}
-                  readOnly
-                />
-                <span>{review.average_rating}</span>
-              </div>
-              <div>{review.comment}</div>
-            </div>
-          </div>
-        ))}
-        <Pagination
-          className="flex justify-end"
-          count={paginationCount}
-          color="primary"
-          onChange={handlePaginationChange}
-        />
-      </div>
-
-      <div>
-        <Dialog
-          open={isPopupPhotoGallery}
-          onClose={handleClose}
-          maxWidth={"md"}
-          className="flex justify-center"
-        >
-          <DialogContent>
-            <StandardImageList
-              itemData={workspaceDetailsData.data.workspace_images}
-            />
-          </DialogContent>
-        </Dialog>
+            </DialogContent>
+          </Dialog>
+        </div>
       </div>
     </div>
   );
